Nieuwe puzzel changes (nog niet af, vooral logic gemaakt), kleine hometown tutorial/aanwijzingen

+++ Assets/Scripts/Management/LevelStarts/Level1Start.cs
@@ -4,10 +4,8 @@ public class Level1Start : MonoBehaviour
 {
     void Start()
     {
-        if (PlayerPrefs.HasKey("Level2Active"))
-        {
-            PlayerPrefs.DeleteKey("Level2Active");
-        }
+        if (PlayerPrefs.HasKey("Level2Active")) { PlayerPrefs.DeleteKey("Level2Active"); }
+        else if (PlayerPrefs.HasKey("Level3Active")) { PlayerPrefs.DeleteKey("Level3Active"); }
 
         PlayerPrefs.SetInt("Level1Active", 1);
         enabled = false;
