+++ Assets/Scripts/Management/LevelStarts/Level2Start.cs
@@ -4,10 +4,8 @@ public class Level2Start : MonoBehaviour
 {
     void Start()
     {
-        if (PlayerPrefs.HasKey("Level1Active"))
-        {
-            PlayerPrefs.DeleteKey("Level1Active");
-        }
+        if (PlayerPrefs.HasKey("Level1Active")) { PlayerPrefs.DeleteKey("Level1Active"); }
+        else if (PlayerPrefs.HasKey("Level3Active")) { PlayerPrefs.DeleteKey("Level3Active"); }
 
         PlayerPrefs.SetInt("Level2Active", 1);
         enabled = false;
